Death Fog => Creeping Death
- Now triggers on kill and targets one enemy unit at a time
